robotik_chess_stack/edge/requirements.txt
pyyaml==6.0.2
requests==2.32.3
pyserial==3.5
opencv-python-headless==4.10.0.84
numpy==1.26.4
